refactor 2nd factor auth
