--- tests/TwoFactorAuthentication.spec.ts
@@ -1,86 +1,0 @@
-import { PlaywrightTestArgs, test } from '@playwright/test';
-import { createRolleAndPersonWithPersonenkontext, UserInfo } from '../base/api/personApi';
-import { getServiceProviderId } from '../base/api/serviceProviderApi';
-import { testschuleName } from '../base/organisation';
-import { typeLehrer } from '../base/rollentypen';
-import { email } from '../base/sp';
-import { DEV, STAGE } from '../base/tags';
-import { deletePersonenBySearchStrings, deleteRolleById } from '../base/testHelperDeleteTestdata';
-import { gotoTargetURL, loginAndNavigateToAdministration } from '../base/testHelperUtils';
-import { generateNachname, generateRolleName, generateVorname } from '../base/utils/generateTestdata';
-import { PersonDetailsViewPage } from '../pages/admin/personen/details/PersonDetailsView.page';
-import { PersonManagementViewPage } from '../pages/admin/personen/PersonManagementView.page';
-import { HeaderPage } from '../pages/components/Header.page';
-
-// The created test data will be deleted in the afterEach block
-let usernames: string[] = [];
-let rolleIds: string[] = [];
-
-test.describe(`Testfälle für TwoFactorAuthentication": Umgebung: ${process.env.ENV}: URL: ${process.env.FRONTEND_URL}:`, () => {
-  test.beforeEach(async ({ page }: PlaywrightTestArgs) => {
-    await test.step(`Login`, async () => {
-      await loginAndNavigateToAdministration(page);
-    });
-  });
-
-  test.afterEach(async ({ page }: PlaywrightTestArgs) => {
-    await test.step(`Testdaten(Benutzer) löschen via API`, async () => {
-      if (usernames.length > 0) {
-        await deletePersonenBySearchStrings(page, usernames);
-        usernames = [];
-      }
-      if (rolleIds.length > 0) {
-        await deleteRolleById(rolleIds, page);
-        rolleIds = [];
-      }
-    });
-
-    await test.step(`Abmelden`, async () => {
-      const header: HeaderPage = new HeaderPage(page);
-      await header.logout();
-    });
-  });
-
-  test(
-    'Prüfen, ob es möglich ist einen Token zurückzusetzen',
-    { tag: [STAGE, DEV] },
-    async ({ page }: PlaywrightTestArgs) => {
-      let userInfoLehrer: UserInfo;
-
-      await test.step(`Testdaten erstellen`, async () => {
-        userInfoLehrer = await createRolleAndPersonWithPersonenkontext(
-          page,
-          testschuleName,
-          typeLehrer,
-          generateNachname(),
-          generateVorname(),
-          [await getServiceProviderId(page, email)],
-          generateRolleName(),
-        );
-        usernames.push(userInfoLehrer.username);
-        rolleIds.push(userInfoLehrer.rolleId);
-      });
-
-      const personManagementView: PersonManagementViewPage = new PersonManagementViewPage(page);
-
-      const personDetailsView: PersonDetailsViewPage = await test.step(`Gesamtübersicht öffnen`, async () => {
-        await gotoTargetURL(page, 'admin/personen');
-        await personManagementView.searchAndOpenGesamtuebersicht(userInfoLehrer.username);
-        return new PersonDetailsViewPage(page);
-      });
-
-      await test.step(`Token einrichten`, async () => {
-        await personDetailsView.addSoftwareToken();
-      });
-
-      await test.step(`2FA Status prüfen dass ein Token eingerichtet ist`, async () => {
-        await personDetailsView.check2FASetup(true);
-      });
-
-      await test.step(`Token zurücksetzen`, async () => {
-        await personDetailsView.resetSoftwareToken();
-        await personDetailsView.check2FASetup(false);
-      });
-    },
-  );
-});
